Rename types.py to messytypes.py to prevent file name clash on sys.path which gives error: AttributeError: 'module' object has no attribute 'ClassType'
