+++ messytables/core.py
@@ -115,7 +115,7 @@ class Cell(object):
 
     def __init__(self, value, column=None, type=None):
         if type is None:
-            from messytables.types import StringType
+            from messytables.messytypes import StringType
             type = StringType()
         self.value = value
         self.column = column
